+++ test/world_context_test.dart
@@ -198,6 +198,38 @@ void main() {
           expect(message.text, 'I love writing tests.');
         },
       );
+      test(
+        '.getAssetStore',
+        () {
+          final sound = CustomSound(
+            assetStore: CustomSoundAssetStore.credits,
+            id: 'whatever',
+          );
+          final world = World();
+          final worldContext = WorldContext(
+            game: Game('Test asset stores'),
+            world: world,
+          );
+          expect(
+            worldContext.getAssetStore(sound).assets,
+            world.creditsAssets,
+          );
+          sound.assetStore = CustomSoundAssetStore.equipment;
+          expect(
+            worldContext.getAssetStore(sound).assets,
+            world.equipmentAssets,
+          );
+          sound.assetStore = CustomSoundAssetStore.interface;
+          expect(
+            worldContext.getAssetStore(sound).assets,
+            world.interfaceSoundsAssets,
+          );
+          sound.assetStore = CustomSoundAssetStore.music;
+          expect(worldContext.getAssetStore(sound).assets, world.musicAssets);
+          sound.assetStore = CustomSoundAssetStore.terrain;
+          expect(worldContext.getAssetStore(sound).assets, world.terrainAssets);
+        },
+      );
     },
   );
 }